Commiting a change to the Maxwell's equations solver which now allows one to pass a parameter "tau" to the solver. By default, this parameter is c, the speed of light, and works exactly as before with no change to the performance of the field solver (tested on weibel in 1x2v and two-stream+weibel in 2x2v, growth rates and nonlinear stage look good). However, if tau is made to be larger than the speed of light, this will add extra diffusion to the Ampere-Maxwell (dE/dt) equation, while reducing the diffusion in the Farady induction (dB/dt) equation (an analogous choice of tau lower than the speed of light has the opposite effect).
In the limit that tau is much larger than the speed of light, this should have the effect of adding significant diffusion to the Ampere-Maxwell equation at the grid scale. Other papers have demonstrated doing so in two-fluid systems allows one to mitigate artificial charge separation when one does not resolve the Debye length. Testing will begin on Stampede with a reconnection simulation which forms double layers upstream of the reconnection layer that sit at the grid scale because we do not resolve the Debye length.

I have also added the lisp files for the higher polynomial orders I've played around with over the past few months. Since these do not affect the size of the directory much, and other people may want to experiment with polynomial orders, especially with the tensor basis, I wanted to add them.

I've also added a few functions to out-scripts which use the gcfac factorization function, which is Maxima's most aggressive factorization routine that I can find. The names of the functions are currently not very good and will undergo revision as I clean up the Vlasov Maxima scripts.