Restructured VmLBO and GkLBO app to make input files simpler. Thanks to PC for making the top-level changes to make this possible.
Now collision table takes a list of species to collide with.
Self-species name must be included in this list for self-collisions.
There should also be a corresponding list of collision frequencies if
using constant collision frequency.

The Regression test gk-lbo/rt-lboCross-1x2v-p1.lua shows how to use
cross-species collisions in GkLBO, but we have not yet tested the
implementation is correct.

I believe the formulas for the 'Greene' option in CrossPrimMom need to
be changed to correct for velocity dimensionality. This change, and
additional options for both VmLBO and GkLBO will follow in subsequent
commits.