Mesh refinement/coarsening tests
Add a unit test for the CartFieldInterpolate updater. It uses the IntegratedDGMoment to check that conservation of x^0, x^1 and x^2 is obeyed when transfering between grids.

Currently this unit test only checks 1D grid transfers, but it checks every combination of (2^a)*(3^b) grids ranging from (2^1)*(3^0) to (2^3)*(3^3). I checked by eye all 451 plots that came out of this and they looked ok. Also the IntegratedDGMoment tests passed, all 2700 of them! Note that for p=1 conservation of x^2 is only approximate, so I had to use assert_close and used a 1% tolerance.

It appears that grid transfers may also work with some powers of 5, but not all. Will add higher dimensional tests next.