FEM interpolation
Add Maxima scripts and kernels for prolongation of an FEM field in the FEM MG Poisson solver (internal cells only, will soon add boundary cells). It is currently restricted to 2^a grids, with 2x grid refinement. It seems challenging to code interpolations in a general sense because a local weak problem needs to be solved to generate the stencil/kernel, but with general grid transfers we don't know a priori the size of the problem or the location of the nodes associated with it.

FEM interpolation is formulated with weak equalities, like in DG. Hopefully this avoids the need for a Galerkin discretization of the coarse grid problem (it did in DG).