FEM coarsening
After much solo deliberation, it seems that the simplest thing to do for now is to formulate the FEM restriction operator by building the prolongation operator and transposing it. Some literature suggests that the coarsening operator should be the tranpose of the prolongation; however if I do that the coarse grid solution seems about one half smaller than the fine-grid solution. Perhaps that is the way it's supposed to be; I will experiment. For now I've settled for introducing an extra factor of 1/2 (not the 1/2^dim that seemed to come about in DG and finite difference methods) so that the coarse and fine grid solutions have the same magnitude.

It's unclear whether this transpose approach will work in general. But the weak equality integrals with unspecified integral limits seem to lead to horribly complex stencils, even for the simple case of 2x coarsening on 2^a grids.