FEM residue
Kernels to compute the residue in the FEM Poisson problem, including boundary kernels. There's no easy way to test these but to run the whole MG solver. Before doing that I need to go back and fix a pontential bug in the smoother kernels (surface term seems to be using phi instead of phi'), and I want to rewrite the smoother Maxima scripts using the extended logical space machinery.

Unfortunately because we need to know whether one of the BCs in corner kernels is Dirichlet in order to enforce that at corner nodes, we cannot avoid the combinatorial explosion of corner kernels with different BCs. Some non-corner boundary kernels could be written as Robin-always kernels, but I think that further complicates the lua. Maybe that will be explored later.