FEM MG infrastructure appears to work in 3D
Still need to carefully check the results/convergence. However I'm not commiting the 3D kernel because it is huge (29 MB!).

In 1D and 2D textbook multigrid efficiency appears to be reached, although it depends some on nu1, nu2.

Also did some reorganization of kernels names, to reduce the amount of code. Unfortunately I think the number of restriction/prolongation kernels will triple because I think those need to be BC-aware, or at least they need to be able to differentiate periodic and Dirichlet BCs.