Update FEM prolongation.
Add an attempt at general FEM prolongation (FEMprolongationGeneral.mac). Unfortunately the general FEM prolongation (without assuming anything about the grids) gave way to truly monstruous kernels (>100k character lines in 2D). So that is not going to work, need to find a better way.

For now, 1D FEM prolongation stencil was hand coded to test a formulation consisting of a combination of even and odd cell contributions. I think this can be Maxima generated, even in higher dimensions.

MG solver seems to give the same electric field as FemPerp for periodic BCs in 1D. The potential is shifted by a constant because we are not enforcing uniqueness.

Also rename residue -> residual everywhere (need to renerate DG kernels).