Adding a small amount of flexibility to the Vlasov equation object, which now allows (in 1x1v) the ability update the force component separately from the streaming component. Vlasov regression tests pass and their run-times are the same, so the additional logic of the flag self._onlyForceUpdate should not affect other Vlasov work.
The use case for this is fairly niche: occasionally it is desirable to force the Vlasov equation stochastically, but the supported time-steppers in Gkeyll are all high-order and one cannot introduce stochastic forcing with high order time-stepping scheme (since by definition, stochastic forcing does not have continuous derivatives in time). By introducing this logic, and a few new computational kernels, a user can do an operator split "by-hand" and treat streaming (and potentially collisions) with a high-order time-stepping scheme, while only performing a first-order forward Euler method for the advection in velocity space via this stochastic force.