Switch ConstDiffusionBCs to use recovery scripts
Update the way BCs were applied for ConstDiffusion terms so they use the recovery scripts. We had first implemented this before the recovery script was created, but now that it is things are simplified greatly.

Tested in 1D with the IncompEuler App. Have not tested higher dim but doesn't seem like anyone is looking to use this right now.