Exciting first pass at using recovery to do the collisionless update in velocity space (instead of penalty fluxes) along with some minor corrections to ConstDiffusion to allow for higher dimensional simulations. Because recovery for hyperbolic equations is more prone to the generation of grid-scale modes, there is a lot of synergy between recent additions to the code to allow for hyperdiffusion, and this recovery based approach for the hyperbolic part of the Vlasov equation.
By default, penalty fluxes will still be used (for now? Maybe everything will be recovery in the future!)

Due to the structure of the streaming term, this term will still be treated with upwinding (ultimately this may be the best thing to do, as then we would only require hyperdiffusion in velocity space instead of configuration space too).

New regression test performs the two-stream instability using the recovery fluxes in velocity space, combined with a small amount of hyperdiffusion (grad^6) for stability.