rename Gyrokinetic -> GyrokineticSimpleHelical
Now the GK kernels in master are referred to as GyrokineticSimpleHelical. This is in preparation for merging GK_gen_geo into master.

All the polyOrder=2 regression tests will fail, because those kernels were old and I just regenerated them. Some polyOrder=1 tests fail too, with small differences (~1e-8) I think because some order of operations changed in the kernels. For some reason just regenerating kernels on different computers changes things like a*(f[1]-f[2]) into a*f[1]-a*f[2] or viceversa, maybe because of different versions of Maxima?

Now there's an optional entry in GkGeometry that controls which version of GK kernels to be used.
geometryType = "SimpleHelical"
will use the kernels currently in master (default option). In the near future we'll use this to request general geometry kernels too (or do that if mapc2p is detected).

Since we are going to have 2 versions of GK kernels I also chose to split the GkSheath stuff into its own header file, and at the same time I arranged the sheath Maxima so we can generate the file with a single Maxima command (before it seemed like people were calling multiple maxima scripts and manually copying the result into the .cpp file).