Bring in SimpleHelical into GenGeo
Objective: we are trying to branch GK so we can run with SimpleHelical and GenGeo kernels. In order to do that I worked in 2 branches:
- mergeMasterAndGKGeo: adds 'SimpleHelical' to GK kernel names in master.
- renameGeoKernels: adds 'GenGeo' to GK kernel names in GK_gen_geo, and brings
  in mergeMasterAndGKGeo changes.

So this commit does the following:

a) Restore the SimpleHelical kernels (and other things like .mac scripts, header
  files, etc) from the mergeMasterAndGKGeo branch.
b) Adapt infrastructure so one can run either SimpleHelical or GenGeo
  calculations. Input files with mapc2p are assumed to be GenGeo simulations.
  Input files with a regular Cartesian grid is assumed to be a SimpleHelical simulation.
  There's a flag that one can pass in GkGeometry, geometryType="SimpleHelical" or "GenGeo", but the behavior when this flag is passed with the wrong grid object may be undesired (it hasn't been tested).
c) Important: with this commit every GK regression test will likely fail. Reasons:
  - Both the SimpleHelical and GenGeo kernels were regenerated (in the
    mergeMasterAndGKGeo and renameGeoKernels branches, respectively) and in doing so the order of operations changed in some places; e.g. a*(f[1]-f[2]) <-> a*f[1]-a*f[2]. I don't understand why this happens, perhaps because different people have different versions of Maxima.
  - In order to read in sources, geometry, ICs, and background distributions we
    are now outputting the ghost cell for those, while master was not doing that. So the reg tests will complain about some files not being the same size.
  - It's possible that some GK reg tests that in master were being run with
    SimpleHelical are now run with mapc2p (GenGeo), e.g. gk-kbm. Perhaps we should restore some of those SimpleHelical tests under a different name, or re-name the ones with GenGeo.
  - Master was projecting some things with ProjectOnBasis (e.g. bmag, bdriftX,
    etc). I think in this commit we are doing the same, but one should check that we are not using EvalOnNodes were master developers had decided it was best to use ProjectOnBasis. Or if it doesn't matter maybe we should just use EvalOnNodes everywhere.
d) Failing SimpleHelical tests: the failing messages either indicate a difference in file size, or small differences. To be extra sure I visually checked that phiSq is the same here to that in the mergeMasterAndGKGeo branch (this should be the same as master +/- small differences).
e) Failing GenGeo tests: same as (d) above, except that the phiSq here was compared against that produced in the GK_gen_geo branch.

If we think this is good, we could proceed as follows to merge things into master:
1. Merge mergeMasterAndGKGeo into master.
2. Merge renameGeoKernels into master.

or alternatively one could try merging renameGeoKernels directly into master. We also have to decide whether we want to merge renameGeoKernels into GK_gen_geo or just leave GK_gen_geo as is for historical record-keeping (not to be modified from here on, and ideally we would work in master or a new, shorter-lived branch).