Modifications to Maxwell's and Vlasov to support upwind fluxes. This switches the Maxwell surf function to instead use the original flux instead of the modified flux and the Vlasov surf function to evaluate the modal basis functions at quadrature points, evaluate the upwind direction, then transform back to the modal basis and evaluate the surface integrals.
This commit also updates a lot of the header files based on recent development, including the change from the keyword restrict to GKYL_RESTRICT, the addition of GKYL_CU_DH decorator for GPU capabilities.

This commit also includes a slight modification to which kernels are generated based on how prohibitively expensive it is to evaluate the upwind fluxes in higher dimensions for p>1. This can be changed but is a nice default for now.