Debugging upwind_quad_to_modal function (making sure all variables are deleted before each new dimension). IMPORTANT: legendre_p has some strange precision issues for the types of operations we want to perform. For example, in 5D with p=1, the coefficients upon integration should be 0.25. Some are 0.25. Others are  9.49900492042053e-9*3.7220045e+7 ~ 0.25.
As such, for right now, I am switching the gaussOrd function (ONLY THE GAUSSORD FUNCTION BECAUSE THIS IS ALL THAT'S NEEDED FOR UPWINDING) to use the analytic formulae for the ordinates for low n. If we can figure out a way to increase the precision of legendre_p that might be all that's needed, but I would prefer to not have this error propagation.