Some clean up based on ongoing tests. Due to a seeming inability to enforce important constraints in p>1 in greater than 1D (such as b . b = 1), abandon rewrite of primitive variable solve using dg_bin_op for now. The inverse operation works well in p=1 in 1D, 2D, and 3D and enforces important constraints. We need to introduce the positivity check in the primitive moment solve to avoid troublesome issues when rho, p_parallel, or p_perp < 0, but that can be handled separately.
Also there were some incorrect comments in the bin_op routines that have been corrected. These were found because in trying to switch the primitive variable solves to dg_bin_op, it was found that it was *easier* to enforce constraints with tensor bases instead of Serendipity basis, but since we could not enforce constraints for all the tests that worked with p=1, we are still abandoning the higher polynomial push for now.