Feature extension for PKPM. Now utilizing b_i b_j at the surfaces to compute a surface expansion of p_ij where it is needed (Lax fluxes and euler update). This feature extension now makes the entirety of PKPM more uniform in its structure. All nonlinear quantities (b, bb, p_ij, the Lax penalization, u, etc.) are computed at the surface with the needed weak operations and then passed to the corresponding surface updates.
There is further exploration that could be done utilizing these surface expansions for the volume update as well (similar to what the entropy stable folks do where Gauss-Lobatto quadrature is utilized for volume updates). But such a construction is somewhat subtle as we did all of these surface expansions to maintain properties of their construction at the surface (i.e. that rhou = rho*u at the surface or b . b = 1 at the surface). Since the properties we worked hard to preserve are not known point-wise there may be no benefit to that construction but further experimentation is needed.

For now, I will move on to clean up and final debugging before the merge.