Commiting a new set of GK kernel generators based on some recent experiments to try to reduce GK cost and improve robustness.
A number of features and extensions are present. The principal change is a change to how the surface fluxes are computed which has a more precise and stronger L2 stability proof with the numerical flux defined as:
F = alpha_surf*(1/2 (f^+ + f^-) - sign(alpha_surf)/2 *(f^+ - f^-))

where the same lessons and experience for transforming between nodal and modal bases are used to construct sign(alpha_surf) as a modal expansion.

We also save on cost by defining a boolean variable which checks if sign(alpha_surf) is single-signed (either +1 or -1) and then just using a traditional upwind flux.

We also move the computation of the surface expansion of alpha to its own function, and in the process we caught a bug in how the surface alpha was being computed (the pre-computation of B*/Bmag could cause alpha to be discontinuous at the surface).

We have also added a new set of kernels with no toroidal field (by=0) which further reduces the cost of the kernels and is useful for both mirror calculations and slab calculations like LAPD relative to the full general geometry case.

The refactor of the surface fluxes has been partially extracted into its own set of functions which should allow this more robust numerical flux to be used with all kinetic equations, Vlasov, SR Vlasov, PKPM, etc. which should also help robustness for these equations systems.