First pass at PKPM kernel refactor that reworks the surface kernels to use the 10 moment Riemann solve with Roe fluxes, supports a characteristic limiter based on that 10 moment Riemann solve with Roe fluxes, and supports an implicit source split similar to the FV scheme that separates the du/dt = E + u x B and dE/dt = J pieces from the hyperbolic component.
Initial results are promising, though energy conservation is slightly worse even as the solutions seems to have gotten much smoother (and this is without limiters). I think the next step is to experiment with the implicit source solve being on du/dt (instead of dJ/dt) to make it easier to eventually handle cross-species collisions, and to change what the quadratic invariant conserved by this operation is (it's not J^2, it is rho u^2). I am not sure if this will work better or worse though at first glance. But I think in general we need to move towards a genuine primitive variable solve since that will be what is required by other systems (such as the relativistic variant of PKPM).