Intermediate commit with refactor of scripts to accommodate nonuniform meshes with magnetic fields (correctly handling the transverse inverse velocity Jacobians when computing grad_v(H) x B).
I have also separated the EB kernels into E kernels and B kernels to facilitate these improvements to functionality, allowing for an easy separation into Bx, By, and Bz kernels (instead of traditionally what we do with grouping by grad_v(w)). In this way, we can include the relevant missing transverse Jacobian factor (since for the magnetic field kernels, we will utilize f, not Jf) and group terms by the component of the field. Since ultimately all of these subtleties only arise in 3V, I have structured the Maxima around this, which makes the Maxima a little more dense in terms of conditionals, but hopefully reasonably readable.

I want to do an experiment where I further break up the lines in magnetic field updates by which component of the field is being multiplies (i.e., one line is all the terms multiplying Bx[0], then a line for Bx[1]). This separation would then allow me to more easily generate the 2x3v p2 tensor kernels for use in the reconnection simulation which, while dense, would hopefully not give the compiler any trouble once the lines are separated in this way.